requirements.txt
# llama-index>=0.9.0
# llama-index-embeddings-gemini>=0.1.0
# llama-index-llms-groq>=0.1.0
# python-dotenv>=0.19.0
# mysql-connector-python>=8.0.0
# pandas>=2.0.0
# numpy>=1.24.0
# plotly>=5.0.0
# chainlit>=0.7.0
# python-dotenv>=0.19.0

# llama-index>=0.9.0
# llama-index-embeddings-gemini>=0.1.0
# llama-index-llms-groq>=0.1.0
# python-dotenv>=0.19.0
# mysql-connector-python>=8.0.0
# pandas>=2.0.0
# numpy>=1.24.0
# plotly>=5.0.0
# chainlit>=0.7.0


chainlit==1.0.0
llama-index==0.9.48.post3
llama-index-llms-groq==0.1.3
mysql-connector-python==8.3.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
python-dotenv==1.0.0
statsmodels==0.14.1